requirement.txt
# python dependencies
functions-framework==3.*
requests==2.20.0
pymongo
datetime
flask
python-dotenv